[RPC Gateway] Reduce DB sync frequency and load (#548)
It's very tricky to directly control the number of DB syncs.
The configs here only control the sync frequency within each lambda.
But if we have way too many lambdas, the number of sync can still be large.
However I still believe this adjustment is helpful when each lambda receives
larger number of requests.

We keep monitoring the DB sync load when we gradually increase the traffic
and launch to more chains. If it seems to go too high, we might need to
drop sync request with a probability.

Current DB read rate 

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/n7CRFgTfi6wAndDXekPd/e9db7a57-edab-4efb-aca2-2c32b290bfb8.png)


Current DB write rate

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/n7CRFgTfi6wAndDXekPd/cd957021-bf15-42e7-8f69-8e17a931e562.png)