fix: deterministic and probabilistic shadow evaluate latency across providers on mainnet (#697)
On mainnet, so far we use the multi-provider probabilistic sampling upon instantiation. We want to change to always instantiate multi-providers. We want to keep other chains as is, because we care about mainnet latency for the most part.

Also we are removing the 10-minute time-based sampling window to allow more deterministic sampling at some probability per quote request. As a result, we add latencyEvaluationSampleProb and healthCheckSampleProb as extra safety measure. The probability was chosen from the previous shadow sampling rate when implementing view-only quoter:
1. mainnet 0.1% https://github.com/Uniswap/routing-api/pull/584
2. arbitrum 0.2% https://github.com/Uniswap/routing-api/pull/599
3. base 0.05% https://github.com/Uniswap/routing-api/pull/598
4. polygon 0.5% https://github.com/Uniswap/routing-api/pull/595
5. optimism 1% https://github.com/Uniswap/routing-api/pull/603
6. bnb 1% https://github.com/Uniswap/routing-api/pull/608
7. celo 10% https://github.com/Uniswap/routing-api/pull/609